Refactor: add schemas directory
